feat: remove cjs files for rollup build
