--- src/index.cjs
@@ -1,1 +1,0 @@
-module.exports = require('./bic-code.cjs')
