--- src/index.esm.js
@@ -1,5 +1,0 @@
-import { validateFullBIC } from './bic-code.esm.js';
-
-export default  {
-    validateFullBIC
-}
