+++ src/index.js
@@ -1,0 +1,5 @@
+import { validateFullBIC } from './bic-code.js';
+
+export default  {
+    validateFullBIC
+}
